Add own copy of linux kernel includes
Otherwise it's impossible to build on anything earlier than RHEL6.
The code is unusable on anything earlier than RHEL6 anyway, but
it is a good thing to be able to compile it, for a number of reasons.

These interfaces are pretty static, so this should not bring us
lots of troubles.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>