Introduce and use SYSEXIT_MKNOD
(1)
create_ploop_dev() was returning -1 which is wrong because this code is used
as a return code from say ploop_mount(). Since there was no appropriate or
close error code let's introduce a new one: SYSEXIT_MKNOD.

(2)
ploop_getdevice() was not checking for a potential error from mknod().
Add a check, log and return proper error.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>