lib/pcopy.c send_process(): if TRACK_INIT fails, do ABORT and retry
In case TRACK_INIT fails with EBUSY, it was not finished (maybe because
the previous instance of ploop copy was killed), so let's do TRACK_ABORT
and retry.

This doesn't handle the case when another instance of ploop copy is
running, but it should be handled by some other means, e.g. file locking.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>