ploop balloon discard: do syncfs() before processing
I have noticed a situation when ploop balloon discard working on a CT fs
in which CT is idle does not discards everything, so when you run it
for the second time, it does some more. So that creates a situation in
which we need to run ploop balloon discard a few times, even if no one
is doing anything with the file system.

After discussion with Maxim and Andrey, we agreed that Most probable fix
is to do syncfs() for mount point before we start the process.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>