create_balloon_file(): umount, rm temp mnt point
In create_balloon_file() which is called by ploop_init_image()
which is called by ploop_create_image(), we create a temporary
directory to be used as mount point, and mount the ploop image
(in order to create balloon file). Neither the mounted image
nor the temp dir is needed after it, so let's roll back properly
and do umount and rmdir.

Notes
1. Not sure why umount was done before only in the error case.
2. Strictly speaking umount is not necessary here because we do it
   later in ploop_umount_image. But it is needed for rmdir to succeed.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>