ploop-copy -s: added mode to copy to local file
ploop copy -s was writing data to stdout (for ploop copy -d
to receive). This patch adds more to write to an opened file.

send_process() new argument is_pipe should be set to non-zero
value if we're writing to a pipe, and to 0 if we're writing to
a local file.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>