Complete current in-kernel maintenance state
There are a few situations (e.g. merge) where user-space can be killed/aborted
but in-kernel ploop remains in some maintenance state.

This patch adds a simple function, which inquires current in-kernel
maintenance state, then complete operation in some way or another,
dependently on this current state.

Added call to
* snapshot	create_snapshot()
* resize	ploop_resize_image()
* merge		merge_image()
* mount		add_deltas()
* umount	ploop_stop()