Introduce ploop_get_info_by_descr()
This is the same as ploop_get_info(), but takes the path to DiskDescriptor.xml
as the first argument, and don't use it if a cached info (.statfs file) is available.

This helps to save some time getting disk usage info for unmounted ploop device.

Old function ploop_get_info() is still available for backward compatibility.
The intention is fix it later, i.e.
- make ploop_get_info() static
- remove its proto from include/libploop.h
- remove an extra call to read_statfs_info() call from ploop_get_info()

Signed-off-by: Kir Kolyshkin <kir@openvz.org>