Replace ploop_{alloc,read}_diskdescriptor() with ploop_read_disk_descr()
Commit c0608b3 introduced read_dd() function to be used by tools,
which simplifies code using it a lot (see also commit 631f468).
This commit does the same for the whole library, changing the external
APT as well.

In fact, outside of the library, there is no use of separate calls to
allocate and read a disk descriptor. Even in the library itself, there
is a single place in ploop_create_image() which calls
alloc_diskdescriptor().

Software using ploop-lib (i.e. vzctl) must be modified appropriately.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>