setver.sh: do set clean=yes automatically
setver.sh have two main use cases
1 set version and then run make dist
2 build (and maybe install) package

In first mode, it just sets the version (which was the intention
of the first setver.sh), and a new version is then used by make dist.
So we need to keep those changes.

In second mode, then the script finishes we already have packages,
so the changes made were already used, and it makes sense to roll
them back. This is what this patch does.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>