ploop_resize_image(): do not return -1
Fix a few cases in ploop_resize_image() in which we return -1.

We should not do this as the value is used as an exit code
from ploop resize command.

This, together with the 10 or so previous patches, should significantly
decrease the probability of doing exit(-1) from ploop resize. I am not
sure though I eliminated all such cases.

Reference: ploop@821373
Signed-off-by: Kir Kolyshkin <kir@openvz.org>