include/dynload.h: add
This file was autogenerated before. We can't do that now when we
are adding more functions to libploop.h, since ordering is important.
So, this commit adds a static version of dynload.h as it appeared
in ploop-1.5 (with slight changes such as added/removed comments
and preprocessor inclusion guard). The next commit will introduce
a way to check if the file is up-to-date.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>