[build] introduce hook to check if dynload.h is up to date
With this patch, make will check if dynload.h is up to date,
and exit with an error if it's not. In such case, you need
to manually add missing functions to dynload.h, near the end
of struct ploop_functions.

The case when we remove an external function is not yet
supported.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>