[build] fix check-api logic
Apparently we need to sort both lists of functions before trying to
run diff on those, otherwise diff complains about function positions,
which is not what we want.

So change the logic of
1. gensym.sh to emit just sorted list of functions
2. include/Makefile's check-api target to generate cur-dynload.h
   as a sorted list of functions.

diff will now ignore the differences in positions.

NEVERTHELESS, new functions should be added at the bottom
(see commit 002b574 for example).

Signed-off-by: Kir Kolyshkin <kir@openvz.org>