MAJOR fix potential NULL in get_dev_by_name()
The code to strip the path from the device name is incorrect,
since in case '/' is not found p becomes null.

One way to fix it would be to add

	else
		p = device;

Instead, let's just use the GNU version of basename which does
exactly what we want.

Note this patch relies on _GNU_SOURCE to be defined during compilation,
and string.h to be included.

Reference: ploop@858760
Signed-off-by: Kir Kolyshkin <kir@openvz.org>