print_output(): generalize print_lsof(), improve
This patch generalizes and simplifies print_lsof() and renames it
to print_output() so now it can print output of any command.

Note that it is no longer printing error message, this should be done by
a caller. It is in fact better because it makes sense to print error
message *after* print_output(), otherwise it can be lost in lots of
output.

Also, there are a few improvements
1. Last \n from the command output is stripped
2. Value of errno is saved and restored
3. Command output is more visible:

  --- somecmd someargs output follows ---
  .......
  ...bla bla...
  .......
  --- somecmd someargs finished ---

Signed-off-by: Kir Kolyshkin <kir@openvz.org>