run_prg(): introduce and use ploop_execvp()
This is the same as libc-provided execvp (as far as I am able
to check with strace), the only difference is we have our own set
of directories to look a binary in, instead of of relying on PATH.

We are a library and should not rely on PATH containing /sbin and
be useful in general.

Alternative approach would be to modify PATH, but this is forbidden for
threaded processes (which we could have), or if search in PATH is
failed, do our own (which is very long and cumbersome).

This is a preparation to fix bug #2595.

https://bugzilla.openvz.org/2595

Reported-by: Thomas Lucke (Lucky) <bugzilla.openvz.org@be.lucky.li>
Signed-off-by: Kir Kolyshkin <kir@openvz.org>