Introduce tmpfiles.d/ploop.conf
In modern distros, /var/lock is a symlink to /run/lock, and /run
is on tmpfs so it's not saved between reboots. Now, in order to
have a consistent directory under it, a tmpfiles.d mechanism is
introduced in most distros running systemd (for example see
https://fedoraproject.org/wiki/Packaging:Tmpfiles.d).

This should fix bug #2493. Note for old distros we still carry
/var/lock/ploop in both install target and spec file.

https://bugzilla.openvz.org/2493

Reported-by: Andrei Vinogradov <spamslepnoga@inbox.ru>
Reported-by: Thomas Lucke (Lucky) <bugzilla.openvz.org@be.lucky.li>
Reported-by: Kalin Bogatzevski <kalin@bul.net>
Signed-off-by: Kir Kolyshkin <kir@openvz.org>