ploop_store_diskdescriptor(): fix image name normalization
In case DiskDescriptor.xml file name doesn't have any path components
(i.e. it is literal "DiskDescriptor.xml"), realpath is not called
and basedir is an empty string. Function normalize_image_name() just
removes the first character from the image name if basedir is empty.
Therefore, image file names are not normalized as they should, instead
bogus file names are produced in the resulting DiskDescriptor.xml.

To fix, check get_basedir result and set it to "./" in case it is empty.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>