tools/ploop: allow T suffix for blockdev size
It is not uncommon to have ploop devices of a few TB size, so let's
allow T suffix for size argument.

Document it in usage and ploop(8) man page.

While at it, change lowercase (kmg) to uppercase (KMG) in suffixes in
usage, since usually lowercase is used for bits and uppercase for bytes.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>