umount.ploop: fixed
Well, this script was not working because "ploop umount -d" expects
device (like /dev/ploop123) but it was called with a partition
(like /dev/ploop123p1). We could have modified ploop umount to strip
off partition name, but it's easier just to let ploop itself care about
the device, calling it with -m $MOUNT_POINT instead.

While at it, remove the check for "not mounted" case -- it is done by
ploop anyway.

https://bugzilla.openvz.org/2788

Reported-by: Thorsten Schifferdecker <curx@openvz.org>
Signed-off-by: Kir Kolyshkin <kir@openvz.org>