umount.ploop: improve mtab handling
The last thing the script does is it removes the mount from /etc/mtab.

First, we need to check if /proc/mtab is a file (not symlink), otherwise
we should not do anything.

Next, it's a bad idea to create a new file and explicitly set its mode.
We could have used chmod --reference but it's easier to use sed -i which
takes care of such things. It also makes code smaller and less
complicated.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>