make_fs(): turn off periodic fs checks
Generally, with journalled file systems, there is no need to do
a periodic filesystem check, since it's always consistent (and
if not, it's dirty and fsck will be performed).

For some reason, though, mkfs still sets these "check after N mounts"
and "check after M days" flags, although I'm pretty sure that
distributions do clear those if you create FS via installer (checked
with CentOS 6). So we need to fix it with tune2fs.

Filesystem will still be checked by ploop on mount if it's dirty
(and the appropriate option (param->fsck) is set), this just removes
periodical checks that are believed to be useless.

https://bugzilla.openvz.org/2769

Reported-by: Stefan Schlesinger <sts@ono.at>
Signed-off-by: Kir Kolyshkin <kir@openvz.org>