ploop_check(): fix cluster value for raw images
In case raw image is checked, we need blocksize to be passed,
as it is needed for check_and_repair_sparse(). Since the latter
expects size in bytes, not sectors, let's convert it the same
way as for non-raw images (see the code "cluster =" in the
function below).

Reference: ploop@943079
Signed-off-by: Kir Kolyshkin <kir@openvz.org>