fsck: use fsck.ext4 not e2fsck
In RHEL 5, e2fsck can only check ext2 and ext3 -- for ext4
a utility called e4fsck is required. But there is no such thing
as e4fsck in RHEL 6.

The easiest solution (suggested by Paparaciz) is to use fsck.ext4,
which should handle ext4, as well as ext3 and ext2.

https://bugzilla.openvz.org/2812#c3

Reported-by: Pavel Odintsov <pavel.odintsov@gmail.com>
Reported-by: Paparaciz <helpaz@gmail.com>
Signed-off-by: Kir Kolyshkin <kir@openvz.org>