ploop convert -f: do not return SYSEXIT_PLOOPFMT
When trying to convert ploop image from v2 to v1, we check for
its size and return an error if it is too big for v1. Exit code
SYSEXIT_PLOOPFMT is used for corrupted ploop images, so
it should not be used in this case. Use SYSEXIT_PARAM instead,
the meaning is one can not specify -v1 format for large ploop images.

Fix to commit 17ebb77/r891129.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>