check_and_repair_sparse(): rm excessive "the"
There is too much "the" in these messages.

> Warning: the extent with unexpected flags

We haven't yet refered to any extent, so there should be no "the".

> Unable to statfs the delta file

One rule is if there's an adjective before noun, "the" should be used
because the adjective makes the noun specified/particular. Following
that rule, it should always be "the delta file". In the meantime, we
have many delta files and no one of them is the one, so drop "the".

Signed-off-by: Kir Kolyshkin <kir@openvz.org>