Makefile.inc: fix dpkg-architecture handling
If dpkg-architecture is not available, DEB_HOST_MULTIARCH will still be
set (to an empty value) so 'ifdef' is always true and we end up with
LIBDIR=/usr/lib/ which is not what we want.

Plus, there can be a case when dpkg-architecture binary is there
but doesn't work as it should (for example, on my Gentoo box with
dpkg ebuild installed).

To fix, let's assume that no output from dpkg-architecture means it's
either absent or failed, and we should try other means of finding out
the proper LIBDIR value.

Alternative, more complicated version is:

 DEB_HOST_MULTIARCH ?= $(shell dpkg-archistecture \
		-qDEB_HOST_MULTIARCH 2>/dev/null || echo Error)
 ifneq "$(DEB_HOST_MULTIARCH)" "Error"

But it looks more complicated and doesn't do any more good.

https://bugzilla.openvz.org/2877

Reported-by: "slepnoga ( Andrei Vinogradov )" <spamslepnoga@inbox.ru>
Signed-off-by: Kir Kolyshkin <kir@openvz.org>