ploop_read_dd(): print error
As per Igor's suggestion, in case any of the required di field
is NULL, print an appropriate error.

Note that ploop_read_dd() is an internal function, and currently
everyone calling it makes sure all these fields are set, so this
check is always false, and the error message added by this patch
is never going to be printed. So this is more like an assert.

Reference: ploop@973782
Signed-off-by: Kir Kolyshkin <kir@openvz.org>