ploop mount: remove -u base feature
The feature of having a special value of 'base' for
'ploop mount -u UUID' was added for debugging, and there is
no use case for it. In addition, it is no longer working
since we switched to new ploop_{open,close}_dd() locking
scheme. So, let's just remove it.

Reference: ploop@976676
Acked-by: Igor Sukhuh <igor@parallels.com>
Signed-off-by: Kir Kolyshkin <kir@openvz.org>