ploop_get_base_delta_uuid(): deprecate
There is no known user of this function outside of the library,
and it is not working properly after the switch to new
ploop_{open,close}_dd() locking scheme since DiskDescriptor.xml
can not be read yet by the time this function is called.

Nevertheless, rather than removing it, let's
1 provide a version that prints an error
2 mark the function as deprecated, so linker will
  emit a warning (or an error, depending on flags)

Similar should be done to the rest of the functions
that are to be deprecated.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>