ploop_replace_image(): fix message about bad level
The thing is, di->nimages can be higher than top_level due to snapshots,
so even if level is less than it, it can still be invalid. Therefore,
here we do only a basic sanity check, the proper check (against
top_level) is to be done later. Because of that we shouldn't tell
anything about upper level limit in the error message.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>