ploop check dd.xml: fix by using check_deltas()
Recently ploop check form with DiskDescriptor.xml was introduced.
Its implementation copied some code from check_deltas() which is
now evolving separately. Another problem is ploop check was checking
all deltas, whether they were used in a current configuration or not.

Fix both issues by switching to check_deltas().

One side effect is we are losing an ability to set various flags.
This is probably the right thing to do, since no dangerous options
can be set, and all the needed options are set by default.
So, modify the usage and the man page accordingly, and issue
a warning if any of the flags are set on the command line.

TODO: add -u UUID option.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>