ploop_check: fix for ro fs case
There is nothing wrong in having non-top-level images on a read only
file system. Unfortunately, since commit f5e47191 this is no longer
possible:

	Error in ploop_check (check.c:224): ploop_check: can't open
	/media/ephemeral0/vz/private/3221/root.hdd/root.hdd:
	Read-only file system

So, ploop mount doesn't work at all.

To fix, let's check if an image is on read-only filesystem, and do
a read-only check in such case. If there are erorrs, they will be
reported anyway.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>