ploop_copy_send(): check ofd == stderr
In case we are going to write data to stderr, we need to disable
logging to stderr. Unfortunately currently there's no way to disable
only stderr, leaving stdout intact, so we disable both. Note that
the last error message is still available via ploop_get_last_error().

Hopefully this is a rare situation when someone is reusing stderr.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>