ploop_copy_*: add feedback_fd
Before now, ploop copy protocol used to be one way, from a sender to a
receiver. While it is perfectly fine for most scenarios, there is one
particular operation that can benefit from the explicit completion
notification -- this is remote fsync() introduced in commit d56654a.

Now, if there is a way to establish either a two-way pipe (say using
socketpair() or ssh port forwarding) or just a second pipe back from
the receiver to the sender (not sure how to do this but there might
be a way), we can finally eliminate that ugly 5s wait for the sending
side sync to happen, instead proceeding as soon as it is done.

Overall effect is to optimize the frozen time during container
live migration. Appropriate support is to be added to ploop tools
and then vzmigrate (unfortunately it requires this newer ploop
to be available on both sides, otherwise we fall back to that 5s wait).

Signed-off-by: Kir Kolyshkin <kir@openvz.org>