ploop_copy_send(): implement async send
As pointed out by a comment (removed by this patch), reads from ploop
delta files are O_DIRECT, as ploop is avoiding using page cache
(to prevent double caching, as caching is done on an inner fs level).

So, O_DIRECT reads are slow, and therefore it is good if we can do those
in parallel with writes. This is what this patch implements.

We use two buffers and a separate thread to send data. First, a buffer
is read into, then we signal a sending thread that there is data to
send, specifying the buffer to send it, and then we go on to the next
read operation, using the other buffer this time. When we signal a
sending thread again.

A pthread mutex and a condidional variable is used for synchronization,
so we don't step onto our own toes.

Some performance data below.

This is how sync read/send looks like without this patch, i.e. when read
and send are done in a syncronous fashion (numbers are relative microsecond
timestamps):

0.0097 PREAD
0.0136 SEND
0.0096 PREAD
0.0156 SEND
0.0097 PREAD
0.0153 SEND
0.0097 PREAD
0.0149 SEND
0.0099 PREAD
0.0139 SEND
0.0096 PREAD
0.0166 SEND
0.0097 PREAD
0.0182 SEND

it takes 0.01 second to read a block and 0.015 seconds to send it, and
since it is done syncronously it takes 0.025 seconds to read and send a
block.

After this patch, send and read are parallel (well, after reading the
initial block), so it saves us MIN(read_time, send_time) multiplied
by the number of iteration. Given the data from the example above,
savings should be about 0.01 seconds per block, or 0.07 seconds total.

Now, 0.07 seconds is imprortant because the last part of ploop copy work
during live migration happens after the container is frozen, so it helps
to improve (minimize) container frozen time.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>