ploop_copy_send_param: remove ofd_is_pipe
There is no need to explicitly specify if a file descriptor
is a pipe/socket or not, as we can just use stat() to find it out.

While at it, move the check for valid fd from plooptool_copy()
to ploop_copy_send(), and make it more strict -- instead of checking
that FD is not a TTY we require a file, a pipe, or a socket.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>