De-introduce SYSEXIT_INUSE
This was introduced by the previous commit 22ff44b (aka ploop@1025736)
for the sole purpose of reporting the "in-use flag is set" condition
from get_image_param_offline(), which, in turn, takes it from
errno value explicitly set to EBUSY by open_delta().
This condition is then checked in the only place --
do_create_snapshot().

Note that SYSEXIT_* codes are part of public API, but this
new value is not actually returned by any of the public API
functions. Therefore, let's apply the Occam razor here.

That abovementioned commit also silently fixes returning -1 from
get_image_param_offline() -- this is good and this patch retains
the fix.

So, this patch reverts the logic of checking errno in
get_image_param_offline(), moving the check directly to
do_create_snapshot().

Signed-off-by: Kir Kolyshkin <kir@openvz.org>