ploop_get_max_size: improve description, move to c
We do not have descriptions of our public functions in libploop.h
file, so let's not start doing it. So far, the best documentation
for such functions is their code itself, so let's keep the description
close to the code.

This is an addition to the previous commit 71be8d5

Cc: Nikolay Shirokovskiy <nshirokovskiy@parallels.com>
Signed-off-by: Kir Kolyshkin <kir@openvz.org>