ploop_check(): open image RO, reopen RW later
Most of the time writing to the image file is not required, so
let's optimize for this case and open the file read only. Later,
if an image has dirty flag set, of CHECK_DETAILED is set, reopen
the file in read/write mode.

This also helps with filesystems on which opening the file read/write
is more expensive than opening it read-only.

Reference: PSBM-26499
References: ploop@1036784, ploop@1036927, ploop@1036926, ploop@1002925, ploop@989937
Signed-off-by: Kir Kolyshkin <kir@openvz.org>