check_and_repair_sparse(): reopen image rw if needed
After the commit r1036784, we can be in the situation when image
is opened as read-only but check_and_repair_sparse() needs to
write to it.

As we don't know beforehand whether we will need to write to an image
file or not, let's fix it in a somewhat kludgy way -- as soon as we
found out we need to write, check opened fd mode and reopen the image
read/write if necessary. Note this is only performed once.

References: #PSBM-20250, #PSBM-26499
Reference: ploop@1036952
Signed-off-by: Kir Kolyshkin <kir@openvz.org>