ploop_is_large_disk_supported(): check ploop module first
A user got the following error:

> # ploop snapshot-switch -u cc0930cb-0e5e-46d8-9f4d-83567ce8fb38 DiskDescriptor.xml
> Opening delta /vz/private/120/root.hdd/root.hdd
> del snapshot {5fbaabe3-6958-40ff-92a7-860e329aab41}
> Adding snapshot {5fbaabe3-6958-40ff-92a7-860e329aab41}
> Storing /vz/private/120/root.hdd/DiskDescriptor.xml.tmp
> Error in do_create_delta (ploop.c:345): Unknown ploop image version: 2

It seems like very old kernel or something else, but the actual reason
is ploop module is not loaded, and it takes a Sherlock Holmes to figure
it out from the provided diagnostics.

So, check first that /sys/module/ploop/parameters is there, and if it's not,
complain once.

Core returned is still the same, as it's supposed to be boolean, but at least
we now have a helpful hint in the output.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>