ploop check dd.xml: skip check if ploop is used
Check should not be done if this set of images is in use by some
ploop device. So, add a call to ploop_get_devs() and return an error
if it finds that the device is in use.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>
Acked-by: Maxim Patlasov <mpatlasov@parallels.com>
(cherry picked from commit a4de0b6405e8c14319758c0f0ad42ff520ad7205)