check_mount_restrictions(): check for all images
Perform checking for all images that will be mounted since
they could be placed on different devices.

[kir@: extended description follows]

check_mount_restrictions() is called before ploop mount to check
that the underlying FS is sufficient (has extents) and is properly
mounted (no "bad" mount options).

Unfortunately we only do it for base delta, which is OK for default
setup, but there might be cases when deltas reside on different
mounts.

This patch adds checking of underlying FS for all images. As such
checking might be expensive we try to optimize for the common case
when all the deltas are on the same FS.

Reference: ploop@1064039