ploop_resize_image(): do not account ext4 metadata to user
When ext4 file system grow, extra ext4 metadata is allocated, and
on a subsequent shrink that data is still accounted to a user,
resulting in reduced available disk space for a user.

It was decided to not account extra ext4 metadata to a user.
The overhead is calculated as n_inodes * inode_size

Reference: PSBM-23635
Reference: ploop@1093781