ploop_trim(): fix len argument
This is a fix to the following error from vzctl compact
(or ploop balloon discard):

	Can't trim file system: Invalid argument

This happens on new kernels (since 042stab10x), that inherited a patch
from RHEL6:

 2.6.32-487.el6 [fs] ext4: Avoid underflow in ext4_trim_fs() (Lukas Czerner)[998785]

That RHEL6 patch adds a check for range.len and returns an error
if it is incorrect. On old kernels this function did nothing but silently
returned success.

ioctl(FITRIM) sets len argument to the number of processed blocks.
If no blocks were processed, len is set to 0 and next ioctl fails
with EINVAL.

Reference: PSBM-30973
Reference: ploop@1096253

Acked-by: Andrey Vagin <avagin@parallels.com>