{mount,umount}.ploop: fix wrt file names with spaces
This patch is needed for ploop mount script helpers to deal
with files containing special characters (like spaces).

Some characters in file names in /etc/mtab and /proc/mounts
are encoded in a special way:

	SPACE	-> \040
	TAB	-> \011
	NEWLINE	-> \012
	\	-> \134

When reading from /proc/mounts, we need to unmangle (decode) these
escape sequences. When writing to /etc/mtab, we need to properly
encode.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>