umount.ploop: fix mtab removal
For some directories, for example ones that contain dashes, our
sed expression is not working, and so /etc/mtab line is not getting
removed.

The fix is to use simpler regex. We should match anything that starts at
the beginning of line, ends with /DiskDescriptor.xml and doesn't contain
spaces.

https://bugzilla.openvz.org/3164

Reported-by: Alexander E. Zabiralov <zabiralov@gmail.com>
Signed-off-by: Kir Kolyshkin <kir@openvz.org>