ploop merge: obsolete form with DD.xml
It's not quite clear how ploop tool ended up with two commands doing the
same thing, but it did:
 - ploop merge <dd.xml>
 - ploop snapshot-merge <dd.xml>
are doing the same thing.

Now, "ploop merge" was originally a low-level ploop command, while
"ploop snapshot-merge" is a form to be used with dd.xml.

Let's make it this way to avoid user confusion and code repetition.

Now, if ploop merge is used with dd.xml, ploop suggests to use
snapshot-merge command. Also, -u and -A arguments are ignored.

Acked-by: Igor Sukhih <igor@parallels.com>
Signed-off-by: Kir Kolyshkin <kir@openvz.org>