print_sys_block_ploop(): print less info
This function is called when EBUSY is returned from ADD_DELTA
or REPLACE_DELTA ploop ioctl, in order to figure out why delta
is busy (presumably it is used by another ploop device).

As we don't need to print the whole contents of /sys/block/ploop*,
but rather some specific info from pdelta. According to Maxim Patlasov,
having pdelta/{image,ro,io} is enough.

This helps to reduce the output, as well as time of command execution
considerably.

Acked-by: Igor Sukhih <igor@parallels.com>
Signed-off-by: Kir Kolyshkin <kir@openvz.org>