ploop replace: implement -k
Currently, replacing a delta means changing its file name. This commit
adds a flag to keep the same file name. If this flag is set, after
successful replace a new delta is renamed to the old one (removing it!).
This makes it possible to replace a delta without modifying dd.xml

Signed-off-by: Kir Kolyshkin <kir@openvz.org>